#1 fix: build match package without the vectors tag via KNN build-tag split
internal/match/match.go called bleve's SearchRequest.AddKNNWithFilter
unconditionally, but bleve v2.6.0 defines that method only under its
//go:build vectors constraint. So the match package (and thus any
tag-free `go build ./...` / `go vet`) failed to compile without the
vectors tag, even though the rest of the tree builds tag-free.

Split the KNN search call into build-tagged helpers, mirroring bleve's
own search_knn.go / search_no_knn.go pattern:

- knn_vectors.go   (//go:build vectors)  runs the real AddKNNWithFilter.
- knn_novectors.go (//go:build !vectors) stubs it, reporting vector
  search unavailable so callers fall back to BM25 text matching.

MatchVector short-circuits with a clear "built without vectors tag"
error when vectorSearchSupported is false. The daemon's resolveSignature
now clears vec on any MatchVector error so the vec==nil BM25 fallback
actually runs — previously the "trying text match" log was empty and
resolution silently dropped to hash-only.

Tests: vector cases moved behind //go:build vectors; text/BM25 cases
stay untagged so they run in both configs. Adds
TestMatchVectorFilteredKNNExcludesNearerOutOfScope, a regression test
that crowds the query with four out-of-scope neighbors (more than
matchK) and one far in-scope row, so a dropped or unfiltered KNN filter
fails loudly.
